backend/requirements/local.txt
-r ./base.txt

Werkzeug==0.14.1  # https://github.com/pallets/werkzeug
ipdb==0.11  # https://github.com/gotcha/ipdb
Sphinx==1.8.4  # https://github.com/sphinx-doc/sphinx
psycopg2==2.7.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.670  # https://github.com/python/mypy
pytest==4.2.0  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.5  # https://github.com/PyCQA/flake8
coverage==4.5.2  # https://github.com/nedbat/coveragepy
pre-commit==1.14.3
black==18.6b4

# Django
# ------------------------------------------------------------------------------
factory-boy==2.11.1  # https://github.com/FactoryBoy/factory_boy

django-extensions==2.1.5  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.4.7  # https://github.com/pytest-dev/pytest-django
django-cors-headers==2.1.0